get events from an artifact set using rcmdcheck::parse_check

get events from an artifact set using rcmdcheck::parse_check

get_events(apathset)

Arguments

apathset
instance of artifact_paths

Examples

setup_demo_artifacts()
eg = select_artifacts(paste(tempdir(), "tokay2", sep="/"),
     c("IRanges", "S4V*", "a4\\."))
get_events(eg)

#> $IRanges
#> $IRanges$warnings
#> [1] "checking whether package 'IRanges' can be installed ... WARNING\nFound the following significant warnings:\n  IRanges_constructor.c:43:6: warning: too many arguments for format [-Wformat-extra-args]\n  IRanges_constructor.c:66:6: warning: too many arguments for format [-Wformat-extra-args]\n  IRanges_constructor.c:86:6: warning: too many arguments for format [-Wformat-extra-args]\n  IRanges_constructor.c:98:5: warning: too many arguments for format [-Wformat-extra-args]\n  NCList.c:674:23: warning: '%d' directive writing between 1 and 11 bytes into a region of size 8 [-Wformat-overflow=]\n  Rd warning: C:/Users/biocbuild/bbs-3.13-bioc/tmpdir/Rtmp0Ev81P/R.INSTALLddc4e68b36/IRanges/man/intra-range-methods.Rd:94: missing link 'RangesList'\n  Rd warning: C:/Users/biocbuild/bbs-3.13-bioc/tmpdir/Rtmp0Ev81P/R.INSTALLddc4e68b36/IRanges/man/intra-range-methods.Rd:102: missing link 'RangesList'\n  Rd warning: C:/Users/biocbuild/bbs-3.13-bioc/tmpdir/Rtmp0Ev81P/R.INSTALLddc4e68b36/IRanges/man/intra-range-methods.Rd:123: missing link 'RangesList'\n  Rd warning: C:/Users/biocbuild/bbs-3.13-bioc/tmpdir/Rtmp0Ev81P/R.INSTALLddc4e68b36/IRanges/man/intra-range-methods.Rd:136: missing link 'RangesList'\n  Rd warning: C:/Users/biocbuild/bbs-3.13-bioc/tmpdir/Rtmp0Ev81P/R.INSTALLddc4e68b36/IRanges/man/intra-range-methods.Rd:146: missing link 'RangesList'\n  Rd warning: C:/Users/biocbuild/bbs-3.13-bioc/tmpdir/Rtmp0Ev81P/R.INSTALLddc4e68b36/IRanges/man/intra-range-methods.Rd:160: missing link 'RangesList'\n  Rd warning: C:/Users/biocbuild/bbs-3.13-bioc/tmpdir/Rtmp0Ev81P/R.INSTALLddc4e68b36/IRanges/man/intra-range-methods.Rd:188: missing link 'RangesList'\n  Rd warning: C:/Users/biocbuild/bbs-3.13-bioc/tmpdir/Rtmp0Ev81P/R.INSTALLddc4e68b36/IRanges/man/intra-range-methods.Rd:297: missing link 'RangesList'\nSee 'C:/Users/biocbuild/bbs-3.13-bioc/meat/IRanges.Rcheck/00install.out' for details."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [2] "checking Rd cross-references ... WARNING\nMissing link or links in documentation object 'IntegerRangesList-class.Rd':\n  '[IRanges]{RangesList}'\n\nMissing link or links in documentation object 'intra-range-methods.Rd':\n  'RangesList'\n\nSee section 'Cross-references' in the 'Writing R Extensions' manual."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [3] "checking for missing documentation entries ... WARNING\nUndocumented code objects:\n  '%poutside%' '%pover%' '%pwithin%' 'cvg' 'heads' 'slidingIRanges'\n  'slidingViews' 'tails' 'windows'\nUndocumented S4 classes:\n  'SimpleRangesList' 'IPosList' 'SimpleIPosList' 'Pos'\n  'CompressedIPosList' 'RangesList' 'BaseManyToManyGrouping' 'Ranges'\n  'CompressedRangesList' 'CompressedIntegerRangesList' 'SimplePosList'\n  'PosList' 'CompressedPosList'\nUndocumented S4 methods:\n  generic 'anyDuplicated' and siglist 'RangesNSBS'\n  generic 'as.factor' and siglist 'ManyToOneGrouping'\n  generic 'as.integer' and siglist 'NormalIRanges'\n  generic 'as.integer' and siglist 'Pos'\n  generic 'as.integer' and siglist 'RangesNSBS'\n  generic 'cbind' and siglist 'Rle'\n  generic 'cbind' and siglist 'RleList'\n  generic 'chartr' and siglist 'ANY,ANY,CharacterList'\n  generic 'chartr' and siglist 'ANY,ANY,RleList'\n  generic 'coerce' and siglist 'ANY,IPosRanges'\n  generic 'coerce' and siglist 'AtomicList,RleViews'\n  generic 'coerce' and siglist 'IPosRanges,CompressedIntegerList'\n  generic 'coerce' and siglist 'IPosRanges,IntegerList'\n  generic 'coerce' and siglist 'IRanges,IPosList'\n  generic 'coerce' and siglist 'IntegerRanges,CompressedIPosList'\n  generic 'coerce' and siglist 'IntegerRanges,IPosList'\n  generic 'coerce' and siglist 'PartitioningByEnd,PartitioningMap'\n  generic 'coerce' and siglist 'factor,ManyToOneGrouping'\n  generic 'commonColnames<-' and siglist 'SplitDataFrameList'\n  generic 'cvg' and siglist 'IntegerRanges'\n  generic 'cvg' and siglist 'IntegerRangesList'\n  generic 'endsWith' and siglist 'CharacterList'\n  generic 'endsWith' and siglist 'RleList'\n  generic 'extractROWS' and siglist 'CompressedList,ANY'\n  generic 'extractROWS' and siglist 'IPos,ANY'\n  generic 'extractROWS' and siglist 'MaskCollection,ANY'\n  generic 'extractROWS' and siglist 'NormalIRanges,ANY'\n  generic 'extractROWS' and siglist 'Partitioning,ANY'\n  generic 'extractROWS' and siglist 'Rle,RangesNSBS'\n  generic 'extractROWS' and siglist 'vector_OR_factor,RangesNSBS'\n  generic 'getListElement' and siglist 'CompressedList'\n  generic 'getListElement' and siglist 'CompressedNormalIRangesList'\n  generic 'getListElement' and siglist 'H2LGrouping'\n  generic 'getListElement' and siglist 'MaskCollection'\n  generic 'getListElement' and siglist 'NCLists'\n  generic 'getListElement' and siglist 'Partitioning'\n  generic 'getListElement' and siglist 'Views'\n  generic 'gsub' and siglist 'ANY,ANY,CharacterList'\n  generic 'gsub' and siglist 'ANY,ANY,RleList'\n  generic 'ifelse2' and siglist 'ANY,List,List'\n  generic 'is.unsorted' and siglist 'CompressedIntegerList'\n  generic 'is.unsorted' and siglist 'CompressedLogicalList'\n  generic 'is.unsorted' and siglist 'CompressedNumericList'\n  generic 'isNormal' and siglist 'IntegerRanges'\n  generic 'length' and siglist 'RangesNSBS'\n  generic 'nchar' and siglist 'CharacterList'\n  generic 'nchar' and siglist 'RleList'\n  generic 'pcompareRecursively' and siglist 'IPosRanges'\n  generic 'range' and siglist 'CompressedRleList'\n  generic 'relist' and siglist 'grouping,missing'\n  generic 'replaceROWS' and siglist 'IRanges'\n  generic 'replaceROWS' and siglist 'NormalIRanges'\n  generic 'slidingWindows' and siglist 'IntegerRanges'\n  generic 'startsWith' and siglist 'CharacterList'\n  generic 'startsWith' and siglist 'RleList'\n  generic 'sub' and siglist 'ANY,ANY,CharacterList'\n  generic 'sub' and siglist 'ANY,ANY,RleList'\n  generic 'tile' and siglist 'IntegerRanges'\n  generic 'tolower' and siglist 'CharacterList'\n  generic 'tolower' and siglist 'RleList'\n  generic 'toupper' and siglist 'CharacterList'\n  generic 'toupper' and siglist 'RleList'\n  generic 'updateObject' and siglist 'CompressedList'\n  generic 'updateObject' and siglist 'IPosRanges'\n  generic 'whichFirstNotNormal' and siglist 'IntegerRanges'\n  generic 'windows' and siglist 'list_OR_List'\nAll user-level objects in a package (including S4 classes and methods)\nshould have documentation entries.\nSee chapter 'Writing R documentation files' in the 'Writing R\nExtensions' manual."
#> [4] "checking Rd \\usage sections ... WARNING\nDocumented arguments not in \\usage in documentation object 'nearest-methods':\n  'hits' '...'\n\nFunctions with \\usage entries need to have the appropriate \\alias\nentries, and all their arguments documented.\nThe \\usage entries must correspond to syntactically valid R code.\nSee chapter 'Writing R documentation files' in the 'Writing R\nExtensions' manual."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> 
#> $IRanges$notes
#> [1] "checking installed package size ... NOTE\n  installed size is  5.1Mb\n  sub-directories of 1Mb or more:\n    R      2.2Mb\n    help   1.8Mb"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#> [2] "checking R code for possible problems ... NOTE\nsetCompressedListSummaryMethod : <anonymous>: no visible binding for\n  global variable 'C_fun'\nsetCompressedListWhichSummaryMethod : def: no visible binding for\n  global variable 'C_fun'\nUndefined global functions or variables:\n  C_fun"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#> [3] "checking compiled code ... NOTE\nNote: information on .o files for i386 is not available\nNote: information on .o files for x64 is not available\nFile 'C:/Users/biocbuild/bbs-3.13-bioc/R/library/IRanges/libs/i386/IRanges.dll':\n  Found 'abort', possibly from 'abort' (C), 'runtime' (Fortran)\nFile 'C:/Users/biocbuild/bbs-3.13-bioc/R/library/IRanges/libs/x64/IRanges.dll':\n  Found 'abort', possibly from 'abort' (C), 'runtime' (Fortran)\n\nCompiled code should not call entry points which might terminate R nor\nwrite to stdout/stderr instead of to the console, nor use Fortran I/O\nnor system RNGs. The detected symbols are linked into the code but\nmight come from libraries and not actually be called.\n\nSee 'Writing portable packages' in the 'Writing R Extensions' manual."
#> 
#> $IRanges$errors
#> character(0)
#> 
#> 
#> $S4Vectors
#> $S4Vectors$warnings
#> [1] "checking whether package 'S4Vectors' can be installed ... WARNING\nFound the following significant warnings:\n  LLint_class.c:358:24: warning: too many arguments for format [-Wformat-extra-args]\nSee 'C:/Users/biocbuild/bbs-3.13-bioc/meat/S4Vectors.Rcheck/00install.out' for details."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [2] "checking Rd cross-references ... WARNING\nMissing link or links in documentation object 'Vector-comparison.Rd':\n  '[IRanges]{IntegerRanges-comparison}'\n\nSee section 'Cross-references' in the 'Writing R Extensions' manual."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [3] "checking for missing documentation entries ... WARNING\nUndocumented code objects:\n  'I' 'head.LLint' 'pcompareRecursively' 'sort.List' 't.HitsList'\n  'tail.LLint' 'unname' 'window.LLint'\nUndocumented S4 methods:\n  generic '!' and siglist 'List'\n  generic '<=' and siglist 'List,List'\n  generic '<=' and siglist 'List,list'\n  generic '<=' and siglist 'list,List'\n  generic '==' and siglist 'List,List'\n  generic '==' and siglist 'List,list'\n  generic '==' and siglist 'list,List'\n  generic '[' and siglist 'LLint'\n  generic '[<-' and siglist 'Rle'\n  generic 'anyNA' and siglist 'List'\n  generic 'by' and siglist 'Vector'\n  generic 'coerce' and siglist 'ANY,DataFrame_OR_NULL'\n  generic 'countMatches' and siglist 'ANY'\n  generic 'do.call' and siglist 'ANY,List'\n  generic 'droplevels' and siglist 'DataFrame'\n  generic 'duplicated' and siglist 'List'\n  generic 'eval' and siglist 'expression,Vector'\n  generic 'eval' and siglist 'language,Vector'\n  generic 'extractCOLS' and siglist 'DataFrame'\n  generic 'extractROWS' and siglist 'DataFrame,ANY'\n  generic 'extractROWS' and siglist 'LLint,ANY'\n  generic 'extractROWS' and siglist 'LLint,NSBS'\n  generic 'extractROWS' and siglist 'LLint,RangeNSBS'\n  generic 'extractROWS' and siglist 'Vector,ANY'\n  generic 'filterRules' and siglist 'FilterResults'\n  generic 'from' and siglist 'HitsList'\n  generic 'getListElement' and siglist 'SimpleList'\n  generic 'grep' and siglist 'ANY,Rle'\n  generic 'grepl' and siglist 'ANY,Rle'\n  generic 'head' and siglist 'LLint'\n  generic 'is.na' and siglist 'List'\n  generic 'is.unsorted' and siglist 'List'\n  generic 'match' and siglist 'List,List'\n  generic 'match' and siglist 'List,Vector'\n  generic 'match' and siglist 'List,list'\n  generic 'match' and siglist 'List,vector'\n  generic 'match' and siglist 'list,List'\n  generic 'max' and siglist 'NSBS'\n  generic 'mergeROWS' and siglist 'ANY'\n  generic 'mergeROWS' and siglist 'DataFrame'\n  generic 'mergeROWS' and siglist 'Vector'\n  generic 'normalizeSingleBracketReplacementValue' and siglist\n    'DataFrame'\n  generic 'normalizeSingleBracketReplacementValue' and siglist 'List'\n  generic 'order' and siglist 'List'\n  generic 'pcompare' and siglist 'List,List'\n  generic 'pcompare' and siglist 'List,list'\n  generic 'pcompare' and siglist 'list,List'\n  generic 'pcompareRecursively' and siglist 'List'\n  generic 'pcompareRecursively' and siglist 'list'\n  generic 'rank' and siglist 'List'\n  generic 'rep.int' and siglist 'LLint'\n  generic 'rep' and siglist 'DataFrame'\n  generic 'replaceCOLS' and siglist 'DataFrame'\n  generic 'replaceROWS' and siglist 'ANY'\n  generic 'replaceROWS' and siglist 'DataFrame'\n  generic 'replaceROWS' and siglist 'Rle'\n  generic 'replaceROWS' and siglist 'Vector'\n  generic 'sameAsPreviousROW' and siglist 'Rle'\n  generic 'sameAsPreviousROW' and siglist 'atomic'\n  generic 'sameAsPreviousROW' and siglist 'complex'\n  generic 'sameAsPreviousROW' and siglist 'integer'\n  generic 'sameAsPreviousROW' and siglist 'numeric'\n  generic 'sort' and siglist 'List'\n  generic 'splitAsList' and siglist 'ANY'\n  generic 'splitAsList' and siglist 'SortedByQueryHits'\n  generic 'summary' and siglist 'FilterResults'\n  generic 't' and siglist 'Pairs'\n  generic 'tail' and siglist 'LLint'\n  generic 'to' and siglist 'HitsList'\n  generic 'unique' and siglist 'List'\n  generic 'unique' and siglist 'SimpleList'\n  generic 'window' and siglist 'LLint'\n  generic 'with' and siglist 'Vector'\n  generic 'xtabs' and siglist 'Vector'\nAll user-level objects in a package (including S4 classes and methods)\nshould have documentation entries.\nSee chapter 'Writing R documentation files' in the 'Writing R\nExtensions' manual."
#> 
#> $S4Vectors$notes
#> [1] "checking dependencies in R code ... NOTE\nUnexported objects imported by ':::' calls:\n  'data.table:::as.data.frame.data.table'\n  'stats:::na.exclude.data.frame' 'stats:::na.omit.data.frame'\n  See the note in ?`:::` about the use of this operator."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> [2] "checking R code for possible problems ... NOTE\nevalSeparately,FilterRules : <anonymous>: no visible global function\n  definition for '.'\nUndefined global functions or variables:\n  ."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [3] "checking compiled code ... NOTE\nNote: information on .o files for i386 is not available\nNote: information on .o files for x64 is not available\nFile 'C:/Users/biocbuild/bbs-3.13-bioc/R/library/S4Vectors/libs/i386/S4Vectors.dll':\n  Found 'abort', possibly from 'abort' (C), 'runtime' (Fortran)\nFile 'C:/Users/biocbuild/bbs-3.13-bioc/R/library/S4Vectors/libs/x64/S4Vectors.dll':\n  Found 'abort', possibly from 'abort' (C), 'runtime' (Fortran)\n\nCompiled code should not call entry points which might terminate R nor\nwrite to stdout/stderr instead of to the console, nor use Fortran I/O\nnor system RNGs. The detected symbols are linked into the code but\nmight come from libraries and not actually be called.\n\nSee 'Writing portable packages' in the 'Writing R Extensions' manual."
#> 
#> $S4Vectors$errors
#> character(0)
#> 
#> 
#> $a4
#> $a4$warnings
#> character(0)
#> 
#> $a4$notes
#> [1] "checking dependencies in R code ... NOTE\nPackages in Depends field not imported from:\n  'a4Base' 'a4Classif' 'a4Core' 'a4Preproc' 'a4Reporting'\n  These packages need to be imported from (in the NAMESPACE file)\n  for when this namespace is loaded but not attached."
#> 
#> $a4$errors
#> character(0)
#> 
#>