get reporting artifacts for a Bioconductor collection (e.g., software, experiment, workflow, ...)

get reporting artifacts for a Bioconductor collection (e.g., software, experiment, workflow, ...)

get_report_tgz_cacheid(
  version = "3.14",
  type = "bioc",
  cache = BiocFileCache::BiocFileCache(),
  url = NULL
)

Arguments

version

character(1) defaults to "3.14"

type

character(1) defaults to 'bioc' which implies 'software'; see Note.

cache

instance of `BiocFileCache::BiocFileCache()`

url

defaults to NULL, if supplied, used to retrieve or cache tgz file

Value

A gzipped tarball is downloaded, informatively renamed, copied to a cache, and the cache reference is returned.

Note

Use bbsBuildArtifacts:::valid_types() to see valid values for `type`. If url is NULL, the report.tgz content will be cached with name formed with type, version, and date of command invocation.

Examples

cururl = paste0("file://", system.file("test_report_3.14_bioc_20211210/test_report.tgz", 
    package="bbsBuildArtifacts"))
id = get_report_tgz_cacheid(url=cururl)
BiocFileCache::bfcquery(BiocFileCache::BiocFileCache(), cururl)
#> # A tibble: 1 × 10
#>   rid    rname  create_time access_time rpath rtype fpath last_modified_t… etag 
#>   <chr>  <chr>  <chr>       <chr>       <chr> <chr> <chr>            <dbl> <chr>
#> 1 BFC304 file:… 2022-01-03… 2022-01-03… /Use… rela… /var…               NA NA   
#> # … with 1 more variable: expires <dbl>