untar the artifact archive and return the path

untar the artifact archive and return the path

path_to_untarred_artifact_folders(
  version = "3.14",
  type = "bioc",
  cache = BiocFileCache::BiocFileCache(),
  destination = tempdir(),
  verbose = TRUE,
  url = NULL,
  destbase = "/report"
)

Arguments

version

character(1) defaults to "3.14"

type

character(1) defaults to 'bioc' which implies 'software'; see Note.

cache

instance of `BiocFileCache::BiocFileCache()`

destination

character(1) path to folder to use, defaults to `tempdir()`

verbose

logical(1) will indicate start and end of untar() process

url

passed to `get_report_tgz_cacheid`

destbase

character(1) name of folder that includes all artifacts, defaults to "report"

Value

character(1) path to folder beneath which all artifacts are found

Note

Typically some additional files exist in addition to the package-related folders at the returned path. `:::non_package_pattern()` can be used to find these.

Examples

cururl = demo_url()
td = tempdir()
path_to_untarred_artifact_folders(url=cururl, destination=td, destbase="test_report")
#> starting untar...
#> done
#> [1] "/var/folders/n4/p9th81md60s8nv12yv40sv8m0000gp/T//RtmpLAsCzL/test_report"